requirements.txt
aiosqlite==0.21.0
mysql-connector-python==9.3.0
typing_extensions==4.14.1
